+++ src/lib/govuk/index.ts
@@ -1,0 +1,11 @@
+export { default as CheckboxGroup } from "./CheckboxGroup.svelte";
+export { default as Checkbox } from "./Checkbox.svelte";
+export { default as DefaultButton } from "./DefaultButton.svelte";
+export { default as ErrorMessage } from "./ErrorMessage.svelte";
+export { default as FormElement } from "./FormElement.svelte";
+export { default as Radio } from "./Radio.svelte";
+export { default as SecondaryButton } from "./SecondaryButton.svelte";
+export { default as Select } from "./Select.svelte";
+export { default as TextArea } from "./TextArea.svelte";
+export { default as TextInput } from "./TextInput.svelte";
+export { default as WarningButton } from "./WarningButton.svelte";
