+++ src/pages/ChooseArea.svelte
@@ -5,14 +5,16 @@
   import { initAll } from "govuk-frontend";
   import { Map } from "maplibre-gl";
   import { onMount } from "svelte";
-  import DefaultButton from "../lib/govuk/DefaultButton.svelte";
-  import ErrorMessage from "../lib/govuk/ErrorMessage.svelte";
-  import FormElement from "../lib/govuk/FormElement.svelte";
-  import Radio from "../lib/govuk/Radio.svelte";
-  import SecondaryButton from "../lib/govuk/SecondaryButton.svelte";
+  import {
+    DefaultButton,
+    ErrorMessage,
+    FormElement,
+    Radio,
+    SecondaryButton,
+  } from "../lib/govuk";
   import "maplibre-gl/dist/maplibre-gl.css";
+  import { FileInput } from "../lib/common";
   import { getAuthoritiesGeoJson } from "../lib/common/data_getter";
-  import FileInput from "../lib/common/FileInput.svelte";
   import About from "../lib/sidebar/About.svelte";
   import { bbox } from "../maplibre_helpers";
   import type { Schema } from "../types";
